requirements.txt
fastmcp>=0.2.0
pydantic>=2.0.0
anthropic>=0.20.0
openai>=1.0.0
google-generativeai>=0.3.0
python-dotenv>=1.0.0
aiohttp>=3.9.0
pyyaml>=6.0.0
